requirements.txt
django
psycopg2-binary
djangorestframework
django-filter
markdown